Straighten out infinite recursion in error handling
It seems like the original intent was just to return error codes
without any additional information. This wasn't possible for Utf8Error
because there was no constructor for this error. Also, add some
"cause" information for these errors so that we can get a little extra
information when we fail.